requirements.txt
selenium==4.31.0
fake-useragent==2.2.0
PyYAML==6.0.2
